Allow Attr to set class attribute as well
The Attr method will now let you set the class of a tag. It also lets
you set multiple classes when passed in as a space separated string.
This also makes working with Spark bindings much easier as it can be used to
get all the remaining attributes and passed directly to Attr instead of
having to iterate over the attributes and conditionally use the AddClass
or AdClasses methods.