Cleaning up API in preparation for 1.0 release
- For the 0.9 release, you can still use the methods that are being replaced
  but they are marked with Obsolete tags
- To use Obsolete methods, you must import the HtmlTags.Extended.Compatibility namespace
  to get access to the old method names implemented as extension methods.
  Alternatively, you could copy CompatibilityExtensions.cs to your own project,
  change the namespace to HtmlTags, and remove the Obsolete attributes.
- All members marked Obsolete will be removed for the 1.0 release

Visible() is now Render() - should help differentiate it from Hide(), which renders, but styles with display:none
Child() is now Append() - more consistent with the corresponding jQuery method
Child<T>() is now Add<T>() - consistent with Add() behavior because the child is returned
AddChildren() is now an overload of Append()

Next is now After() - changed to method to allow chaining. Name is consistent with jQuery
EncodeInnerText is now Encoded() - changed to method to allow chaining
UnEncoded() is now an extension method in HtmlTags.Extended.Attributes. Can use Encoded(false) instead.

Many higher-level extension methods were moved to HtmlTags.Extended.Attributes so that they are not imported by default.
Tag building extension methods were moved to HtmlTags.Extended.TagBuilders for the same reason.

Add(string) and Append(string) now accept '>' as a name delimeter (in addition to the original '/'), to be consistent with jQuery