MetaData() now stores data in a dedicated 'data-:' attribute, instead of the 'class' attribute.
The default behavior of the jquery.metadata plugin is to load the metadata from the class attribute.
However, it was discovered that some browsers actually interpret embedded metadata as classes, which is really bad.
To avoid any unintended consequences, metadata should be stored in its own dedicated attribute.

For the metadata plugin to work correctly with tags created by HtmlTags,
you will need to add the following code in a script/page AFTER the jquery.metadata.js script is loaded:

if ($.metadata) {
    $.metadata.setType('attr', 'data-:');
}